Adding tests for PPO indicator

+++ src/test/java/eu/verdelhan/ta4j/indicators/helpers/AverageTrueRangeIndicatorTest.java
@@ -22,7 +22,6 @@
  */
 package eu.verdelhan.ta4j.indicators.helpers;
 
-import eu.verdelhan.ta4j.indicators.helpers.AverageTrueRangeIndicator;
 import eu.verdelhan.ta4j.Tick;
 import eu.verdelhan.ta4j.mocks.MockTick;
 import eu.verdelhan.ta4j.mocks.MockTimeSeries;
@@ -31,8 +30,8 @@ import java.util.List;
 import static org.assertj.core.api.Assertions.*;
 import org.junit.Test;
 
-
 public class AverageTrueRangeIndicatorTest {
+
 	@Test
 	public void testGetValue() {
 		List<Tick> ticks = new ArrayList<Tick>();
@@ -43,11 +42,11 @@ public class AverageTrueRangeIndicatorTest {
 		ticks.add(new MockTick(0, 0, 0, 2));
 		AverageTrueRangeIndicator atr = new AverageTrueRangeIndicator(new MockTimeSeries(ticks), 3);
 		
-		assertThat(atr.getValue(0)).isEqualTo((double) 1d);
-		assertThat(atr.getValue(1)).isEqualTo((double) 2d/3 + 6d/3);
-		assertThat(atr.getValue(2)).isEqualTo((double) (2d/3 + 6d/3) * 2d/3 + 9d/3);
-		assertThat(atr.getValue(3)).isEqualTo((double) ((2d/3 + 6d/3) * 2d/3 + 9d/3) * 2d/3 + 3d/3);
-		assertThat(atr.getValue(4)).isEqualTo((double) (((2d/3 + 6d/3) * 2d/3 + 9d/3) * 2d/3 + 3d/3) * 2d/3 + 15d/3);
+		assertThat(atr.getValue(0)).isEqualTo(1d);
+		assertThat(atr.getValue(1)).isEqualTo(2d/3 + 6d/3);
+		assertThat(atr.getValue(2)).isEqualTo((2d/3 + 6d/3) * 2d/3 + 9d/3);
+		assertThat(atr.getValue(3)).isEqualTo(((2d/3 + 6d/3) * 2d/3 + 9d/3) * 2d/3 + 3d/3);
+		assertThat(atr.getValue(4)).isEqualTo((((2d/3 + 6d/3) * 2d/3 + 9d/3) * 2d/3 + 3d/3) * 2d/3 + 15d/3);
 		
 	}
 }
